backend/requirements.txt
fastapi
uvicorn
groq
pdfplumber
python-dotenv
python-multipart